requirements.txt
# Web framework
Flask==3.0.3

# OpenAI SDK for Azure OpenAI API
openai>=1.3.0

# CORS support for Flask
Flask-Cors==4.0.1

# Load environment variables from .env file
python-dotenv==1.0.1
